about page editor working correctly
